// src/client/io/ClientProt.ts
// GENERATED by tools/gen-prot.ts from the engine. Do not edit by hand.

export const enum ClientProt {
    NO_TIMEOUT = 181,
    IDLE_TIMER = 145,
    EVENT_MOUSE_CLICK = 224,
    EVENT_MOUSE_MOVE = 229,
    EVENT_APPLET_FOCUS = 149,
    EVENT_CAMERA_POSITION = 193,
    ANTICHEAT_OPLOGIC1 = 195,
    ANTICHEAT_OPLOGIC2 = 81,
    ANTICHEAT_OPLOGIC3 = 122,
    ANTICHEAT_OPLOGIC4 = 49,
    ANTICHEAT_OPLOGIC5 = 46,
    ANTICHEAT_OPLOGIC6 = 73,
    ANTICHEAT_OPLOGIC7 = 133,
    ANTICHEAT_OPLOGIC8 = 168,
    ANTICHEAT_OPLOGIC9 = 88,
    ANTICHEAT_CYCLELOGIC1 = 130,
    ANTICHEAT_CYCLELOGIC2 = 154,
    ANTICHEAT_CYCLELOGIC3 = 125,
    ANTICHEAT_CYCLELOGIC4 = 137,
    ANTICHEAT_CYCLELOGIC5 = 85,
    ANTICHEAT_CYCLELOGIC6 = 255,
    ANTICHEAT_CYCLELOGIC7 = 232,
    OPOBJ1 = 97,
    OPOBJ2 = 4,
    OPOBJ3 = 110,
    OPOBJ4 = 147,
    OPOBJ5 = 22,
    OPOBJT = 241,
    OPOBJU = 55,
    OPNPC1 = 252,
    OPNPC2 = 21,
    OPNPC3 = 178,
    OPNPC4 = 30,
    OPNPC5 = 247,
    OPNPCT = 108,
    OPNPCU = 160,
    OPLOC1 = 10,
    OPLOC2 = 45,
    OPLOC3 = 196,
    OPLOC4 = 53,
    OPLOC5 = 126,
    OPLOCT = 218,
    OPLOCU = 184,
    OPPLAYER1 = 220,
    OPPLAYER2 = 51,
    OPPLAYER3 = 13,
    OPPLAYER4 = 189,
    OPPLAYER5 = 69,
    OPPLAYERT = 138,
    OPPLAYERU = 16,
    OPHELD1 = 76,
    OPHELD2 = 177,
    OPHELD3 = 40,
    OPHELD4 = 191,
    OPHELD5 = 79,
    OPHELDT = 112,
    OPHELDU = 200,
    INV_BUTTON1 = 44,
    INV_BUTTON2 = 111,
    INV_BUTTON3 = 124,
    INV_BUTTON4 = 248,
    INV_BUTTON5 = 227,
    IF_BUTTON = 86,
    RESUME_PAUSEBUTTON = 166,
    CLOSE_MODAL = 93,
    RESUME_P_COUNTDIALOG = 180,
    TUT_CLICKSIDE = 146,
    MAP_BUILD_COMPLETE = 214,
    MOVE_OPCLICK = 67,
    SEND_SNAPSHOT = 94,
    MOVE_MINIMAPCLICK = 236,
    INV_BUTTOND = 253,
    IGNORELIST_DEL = 251,
    IGNORELIST_ADD = 192,
    IDK_SAVEDESIGN = 27,
    CHAT_SETMODE = 161,
    MESSAGE_PRIVATE = 107,
    FRIENDLIST_DEL = 203,
    FRIENDLIST_ADD = 235,
    CLIENT_CHEAT = 34,
    MESSAGE_PUBLIC = 156,
    MOVE_GAMECLICK = 234,
}